1. Se agrega comandos en el archivo make para ejecutar las shell 2. Se mueve los modelos a el aplicación CORE 3. Se hiso lectura del excel para comparar y se instancia la clase    ExtraccionArgentina para hacer pruebas 4. Se agrego la libreria OPENXL
