+++ README.md
@@ -16,3 +16,7 @@ se puede visualizar como esta distribuido por etapas el proyecto, en cada etapa 
 # Comandos
 
 `docker-compose run --rm proyectofinalekchy sh -c "python manage.py startapp user"`
+
+## Ejecutar un script python en el contexto de django
+1. `make docker-shell`
+2. `python manage.py shell < dataextraction/lectura_excel.py`
