+++ requirements.dev.txt
@@ -30,3 +30,4 @@ pytz==2022.1
 PyYAML==6.0
 sqlparse==0.4.2
 uritemplate==4.1.1
+openpyxl==3.0.10
