+++ requirements.txt
@@ -5,3 +5,4 @@ pdfplumber==0.7.4
 djangorestframework>=3.12.4,<3.13
 psycopg2>=2.8.6,<2.9
 drf-spectacular>=0.15.1,<0.16
+openpyxl==3.0.10
